Move common code to add missions to the base class
